requirements.txt
aiohttp
asyncpg
discord.py
pydantic
python-dotenv
openai
beautifulsoup4
PyGithub